Increment the version in one place.
Support incrementing the version in one place by allowing the nuspec to
pick up the version from the assembly info, and letting AppVeyor patch
the assembly info as part of the build process.

Add appveyor.yml to the project root. SemVer style version changes
should be made the the core_version property in the appveyor.yml.