Allow migrations to produce additional commands
Alter.Collection can be used now to add additional ICommandData to the
Alter.Collection batch. This makes it so you can safely add new
documents or make changes to additional documents and have it be part of
the same batch / transaction.

This changes the API of Alter.Colleciton though, so this might need to
be changed to allow for both the Action and the Func.