Release version 2.0.0
Bump appveyor.yml `core_version` to 2.0.0. Move x.x.x notes to 2.0.0 in
`CHANGELOG.md`.

Changes in this release:
- Changed (breaking) - The way the MigrationDocument's Id is determined.
Multiple underscores
will now combined into one separator to be used between each section of
the Id.
This could be a breaking change if there are migrations named with
multiple
underscores, depending on the version of Raven client and server in use.
Some
**This could be a breaking change if there are migrations named with
multiple
underscores**, depending on the version of Raven client and server in
use. Some
versions of Raven ignore multiple separators and others do not. This
change
could cause migrations to be run multiple times in some cases if the ids
are not
**could cause migrations to be run multiple times** in some cases if the
ids are not
changed first.
- Added - Ability to inherit from `MigrationAttribute` to specify custom
migration
versions.